[CL-1576] Extract get project id logic into seperate file
