+++ front/app/modules/commercial/matomo/actions.ts
@@ -1,19 +1,14 @@
-// services
-import { projectBySlugStream, IProject } from 'services/projects';
-import { ideaBySlugStream, IIdea } from 'services/ideas';
-
 // routes
 import createRoutes from 'routes';
 import matchPath, { getAllPathsFromRoutes } from './matchPath';
 
 // utils
-import { isNilOrError, NilOrError } from 'utils/helperUtils';
+import { isNilOrError } from 'utils/helperUtils';
 import { tenantInfo, IEvent } from 'utils/analytics';
 import { getUrlLocale } from 'services/locale';
-import { slugRegEx } from 'utils/textUtils';
+import { getProjectId } from './getProjectId';
 
 // typings
-import { Subscription } from 'rxjs';
 import { IAppConfiguration } from 'services/appConfiguration';
 
 export const trackEvent = (
@@ -51,32 +46,12 @@ export const trackPageChange = async (path: string) => {
     window._paq.push(['setCustomDimension', 3, locale]);
   }
 
-  // Update project id custom dimension if project page
-  if (isProjectPage(path)) {
-    const slug = extractProjectSlug(path)
-    if (!slug) return;
+  // Update project id custom dimension if URL is associated
+  // with project
+  const projectId = await getProjectId(path);
 
-    const projectId = await getProjectIdFromProjectSlug(slug);
-    projectSubscriptions[slug].unsubscribe();
-    delete projectSubscriptions[slug]
-    
+  if (projectId) {
     window._paq.push(['setCustomDimension', 4, projectId])
-  } else {
-    window._paq.push(['setCustomDimension', 4, undefined])
-  }
-
-  // Update project id custom dimension if idea page
-  if (isIdeaPage(path)) {
-    const slug = extractIdeaSlug(path)
-    if (!slug) return;
-
-    const projectId = await getProjectIdFromIdeaSlug(slug);
-    ideaSubscriptions[slug].unsubscribe();
-    delete ideaSubscriptions[slug]
-    
-    window._paq.push(['setCustomDimension', 4, projectId])
-  } else {
-    window._paq.push(['setCustomDimension', 4, undefined])
   }
 
   // Set custom URL (override default behavior)
@@ -94,61 +69,3 @@ export const trackPageChange = async (path: string) => {
     window._paq.push(['trackPageView']);
   }
 };
-
-const slugRegExSource = slugRegEx.source.slice(1, slugRegEx.source.length - 2)
-
-const projectPageDetectRegex = RegExp(`\/projects\/(${slugRegExSource})`);
-const projectPageExtractRegex = /\/projects\/([^\s!?\/.*#|]+)/;
-
-const isProjectPage = (path: string) => {
-  return projectPageDetectRegex.test(path);
-}
-
-const extractProjectSlug = (path: string) => {
-  const matches = path.match(projectPageExtractRegex);
-  return matches && matches[1];
-}
-
-const projectSubscriptions: Record<string, Subscription> = {};
-
-const getProjectIdFromProjectSlug = (slug: string) => {
-  return new Promise((resolve, reject) => {
-    const observable = projectBySlugStream(slug).observable
-
-    projectSubscriptions[slug] = observable.subscribe((project: IProject | NilOrError) => {
-      if (isNilOrError(project)) {
-        reject(project)
-      } else {
-        resolve(project.data.id)
-      }
-    })
-  })
-}
-
-const ideaPageDetectRegex = RegExp(`\/ideas\/(${slugRegExSource})`);
-const ideaPageExtractRegex = /\/ideas\/([^\s!?\/.*#|]+)/;
-
-const isIdeaPage = (path: string) => {
-  return ideaPageDetectRegex.test(path);
-}
-
-const extractIdeaSlug = (path: string) => {
-  const matches = path.match(ideaPageExtractRegex);
-  return matches && matches[1];
-}
-
-const ideaSubscriptions: Record<string, Subscription> = {};
-
-const getProjectIdFromIdeaSlug = (slug: string) => {
-  return new Promise((resolve, reject) => {
-    const observable = ideaBySlugStream(slug).observable
-
-    ideaSubscriptions[slug] = observable.subscribe((idea: IIdea | NilOrError) => {
-      if (isNilOrError(idea)) {
-        reject(idea)
-      } else {
-        resolve(idea.data.relationships.project.data.id)
-      }
-    })
-  })
-}
